+++ src/main.js
@@ -1,5 +1,6 @@
 import normalizeException from 'normalize-exception'
 
+import { getColors } from './colors.js'
 import { exitProcess } from './exit.js'
 import { getOpts } from './options/main.js'
 import { printError } from './print/main.js'
@@ -13,6 +14,7 @@ export default function handleCliError(error, opts) {
     error: errorB,
     opts: { silent, stack, props, colors, exitCode, timeout },
   } = getOpts(opts, errorA)
-  printError({ error: errorB, silent, stack, props, colors })
+  const { useColors } = getColors(colors)
+  printError({ error: errorB, silent, stack, props, useColors })
   exitProcess(exitCode, timeout)
 }