+++ src/options/main.js
@@ -37,7 +37,6 @@ const INVALID_OPTS = {
   silent: false,
   stack: true,
   props: true,
-  colors: true,
   exitCode: INVALID_OPTS_EXIT_CODE,
   timeout: DEFAULT_TIMEOUT,
 }