+++ src/print/main.js
@@ -14,20 +14,29 @@ import {
 // Otherwise, we pass the `error` instance to `console.error()`, so it prints
 // not only its `message` and `stack` but also its properties, `cause`,
 // aggregate `errors`, add colors, inline preview, etc. using `util.inspect()`.
-export const printError = function ({ error, silent, stack, props, colors }) {
+export const printError = function ({
+  error,
+  silent,
+  stack,
+  props,
+  useColors,
+}) {
   if (silent) {
     return
   }
 
-  const errorString = serializeError({ error, stack, props, colors })
+  const errorString = serializeError({ error, stack, props, useColors })
   // eslint-disable-next-line no-restricted-globals, no-console
   console.error(errorString)
 }
 
-const serializeError = function ({ error, stack, props, colors }) {
+const serializeError = function ({ error, stack, props, useColors }) {
   const errorA = omitProps(error, props)
   omitStack(errorA, stack)
-  const errorString = inspect(errorA, { colors, depth: PRINT_MAX_DEPTH })
+  const errorString = inspect(errorA, {
+    colors: useColors,
+    depth: PRINT_MAX_DEPTH,
+  })
   restoreStack(errorA, stack)
   const errorStringA = omitStackBracket(errorString, stack)
   return errorStringA