+++ test/print/main.js
@@ -1,6 +1,4 @@
 import test from 'ava'
-import hasAnsi from 'has-ansi'
-import { each } from 'test-each'
 
 import { handleError } from '../helpers/main.js'
 
@@ -8,13 +6,4 @@ const testError = new TypeError('test')
 
 test.serial('Does not log if "silent" is true', (t) => {
   t.is(handleError(testError, { silent: true }).consoleArg, undefined)
-})
-
-each([true, false, undefined], ({ title }, colors) => {
-  test.serial(`Add colors unless "colors" is false | ${title}`, (t) => {
-    t.is(
-      hasAnsi(handleError(testError, { colors }).consoleArg),
-      colors !== false,
-    )
-  })
 })